+++ app/src/main/java/net/gsantner/opoc/ui/SearchOrCustomTextDialog.java
@@ -15,14 +15,12 @@ import android.content.Context;
 import android.content.res.ColorStateList;
 import android.graphics.Color;
 import android.graphics.Typeface;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.support.annotation.ColorInt;
 import android.support.annotation.LayoutRes;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.StringRes;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.AppCompatEditText;
 import android.text.Editable;
@@ -45,23 +43,14 @@ import android.widget.LinearLayout;
 import android.widget.ListView;
 import android.widget.TextView;
 
-import net.gsantner.opoc.util.ActivityUtils;
 import net.gsantner.opoc.util.Callback;
 import net.gsantner.opoc.util.ContextUtils;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.InputStreamReader;
-import java.lang.ref.WeakReference;
 import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Locale;
-import java.util.Queue;
 import java.util.regex.Pattern;
 
-@SuppressWarnings("WeakerAccess")
 public class SearchOrCustomTextDialog {
 
     public static class DialogOptions {
@@ -316,235 +305,6 @@ public class SearchOrCustomTextDialog {
         }
         if (dopt.defaultText != null) {
             listAdapter.getFilter().filter(searchEditText.getText());
-        }
-    }
-
-
-    public static SearchFilesTask recursiveFileSearch(Activity activity, File searchDir, String query, boolean isSearchInFiles, boolean isShowResultOnCancel, Integer maxSearchDepth, Callback.a1<List<String>> callback) {
-        query = query.replaceAll("(?<![.])[*]", ".*");
-        SearchFilesTask task = new SearchFilesTask(activity, searchDir, query, callback, query.startsWith("^") || query.contains("*"), isSearchInFiles, isShowResultOnCancel, maxSearchDepth);
-        task.execute();
-        return task;
-    }
-
-    public static class SearchFilesTask extends AsyncTask<Void, Integer, List<String>> {
-        private final Callback.a1<List<String>> _callback;
-        private final File _searchDir;
-        private final String _query;
-        private final boolean _isRegex;
-        private final boolean _isSearchInFiles;
-        private final boolean _isShowResultOnCancel;
-        private final Integer _maxSearchDepth;
-        private final WeakReference<Activity> _activityRef;
-
-        private final Pattern _regex;
-        private Snackbar _snackBar;
-        private Integer _countCheckedFiles;
-        private Integer _queueLength;
-        private boolean _isCanceled;
-        private Integer _searchDepth;
-        private List<String> _result;
-
-        public SearchFilesTask(Activity activity, File searchDir, String query, Callback.a1<List<String>> callback, boolean isRegex, boolean isSearchInFiles, boolean isShowResultOnCancel, Integer maxSearchDepth) {
-            _searchDir = searchDir;
-            _query = isRegex ? query : query.toLowerCase();
-            _callback = callback;
-            _isRegex = isRegex;
-            _isSearchInFiles = isSearchInFiles;
-            _regex = isRegex ? Pattern.compile(_query) : null;
-            _activityRef = new WeakReference<>(activity);
-            _countCheckedFiles = 0;
-            _isShowResultOnCancel = isShowResultOnCancel;
-            _isCanceled = false;
-            _result = new ArrayList<>();
-            _searchDepth = 0;
-            _maxSearchDepth = maxSearchDepth;
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            if (_activityRef.get() != null) {
-                _snackBar = Snackbar.make(_activityRef.get().findViewById(android.R.id.content), _query + "...", Snackbar.LENGTH_INDEFINITE);
-                _snackBar.setAction(android.R.string.cancel, (v) -> {
-                    _snackBar.dismiss();
-                    preCancel();
-                }).show();
-            }
-        }
-
-        private void preCancel(){
-            if(_isShowResultOnCancel){
-                _isCanceled = true;
-                return;
-            }
-
-            cancel(true);
-        }
-
-        private void getFilesByEqualsFileNames(File currentDir){
-            try {
-                if(!currentDir.canRead() || currentDir.isFile()){
-                    return;
-                }
-
-                File[] files = currentDir.listFiles();
-
-                for (int i = 0; i < files.length; i++) {
-                    if(isCancelled() || _isCanceled){
-                        break;
-                    }
-
-                    _countCheckedFiles++;
-                    File file = files[i];
-                    String fileName = file.getName();
-                    boolean isMatch = _isRegex ? _regex.matcher(fileName).matches() : fileName.toLowerCase().contains(_query);
-
-                    if(isMatch){
-                        String path = file.getCanonicalPath().replace(_searchDir.getCanonicalPath() + "/", "");
-                        _result.add(path);
-                    };
-                }
-            } catch (Exception ex){
-                ;
-            }
-        }
-
-        private boolean isFileContainSearchQuery(File file){
-            boolean ret = false;
-
-            if(!file.canRead() || file.isDirectory()){
-                return ret;
-            }
-
-            try(BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream(file)))) {
-                for(String line; (line = br.readLine()) != null; ) {
-                    if(isCancelled() || _isCanceled){
-                        break;
-                    }
-                    boolean isMatch = _isRegex ? _regex.matcher(line).matches() : line.toLowerCase().contains(_query);
-                    if(isMatch){
-                        ret = true;
-                        break;
-                    }
-                }
-            }
-            catch (Exception ex){
-                ;
-            }
-
-            return ret;
-        }
-
-        private void getFilesByContext(File currentDir){
-            try{
-                if(!currentDir.canRead()){
-                    return;
-                }
-
-                File[] files = currentDir.listFiles();
-
-                for (int i = 0; i < files.length; i++) {
-                    if(isCancelled() || _isCanceled){
-                        break;
-                    }
-
-                    File file = files[i];
-                    _countCheckedFiles++;
-                    publishProgress(_queueLength, _searchDepth, _result.size(), _countCheckedFiles);
-                    if(!file.canRead() || file.isDirectory()){
-                        continue;
-                    }
-
-                    if(isFileContainSearchQuery(file)){
-                        String path = file.getCanonicalPath().replace(_searchDir.getCanonicalPath() + "/", "");
-                        _result.add(path);
-                    }
-                }
-            } catch (Exception ex){
-                ;
-            }
-        }
-
-        @Override
-        protected List<String> doInBackground(Void... voidp) {
-            Queue<File> queue = new LinkedList<File>();
-            queue.add(_searchDir);
-
-            while (!queue.isEmpty() && !isCancelled() && !_isCanceled && _searchDepth <= _maxSearchDepth) {
-                _searchDepth++;
-                _queueLength = queue.size();
-                publishProgress(_queueLength, _searchDepth, _result.size(), _countCheckedFiles);
-                File currentDirectory = queue.remove();
-
-
-                if(!currentDirectory.canRead() || currentDirectory.isFile()){
-                    continue;
-                }
-
-                getFilesByEqualsFileNames(currentDirectory);
-
-                if(_isSearchInFiles){
-                    getFilesByContext(currentDirectory);
-                }
-
-                try{
-                    File[] files = currentDirectory.listFiles();
-                    for(int i = 0; i < files.length; i++){
-                        File file = files[i];
-
-                        if(!file.canRead() || file.isFile()){
-                            continue;
-                        }
-
-                        // ignore symbolic links
-                        File parentFile =  file.getCanonicalFile().getParentFile();
-                        if(parentFile == null || !currentDirectory.getCanonicalPath().equals(parentFile.getCanonicalPath())){
-                            continue;
-                        }
-
-                        queue.add(file);
-                    }
-
-                    _queueLength = queue.size();
-                    publishProgress(_queueLength, _searchDepth, _result.size(), _countCheckedFiles);
-                } catch (Exception ex){
-                    ;
-                }
-            }
-
-            if(_isCanceled && _result.size() == 0){
-                cancel(true);
-            }
-
-            return _result;
-        }
-
-        @Override
-        protected void onProgressUpdate(Integer... values) {
-            super.onProgressUpdate(values);
-
-            Integer queueLength = values[0];
-            Integer queueDepth = values[1];
-            Integer filesFound = values[2];
-            Integer countCheckedFiles = values[3];
-            String snackBarText = _query + "...(" + filesFound + ") qu:" + queueLength + "|" + queueDepth + " c:" + countCheckedFiles;
-            _snackBar.setText(snackBarText);
-        }
-
-        @Override
-        protected void onPostExecute(List<String> ret) {
-            super.onPostExecute(ret);
-            if (_snackBar != null) {
-                _snackBar.dismiss();
-            }
-            if (_callback != null) {
-                try {
-                    _callback.callback(ret);
-                } catch (Exception ignored) {
-                }
-            }
-            new ActivityUtils(_activityRef.get()).hideSoftKeyboard().freeContextRef();
         }
     }
 }
